Enable ktlint, and add detekt gradle plugin
This commit enables ktlint for code formatting and adds the detekt Gradle plugin for static code analysis.
It also updates `libs.versions.toml` to include the detekt version and applies the detekt plugin to the root and app `build.gradle.kts` files.
Additionally it adds `.editorconfig` file to set `ktlint_standard_function-naming` config.
The project `vcs` is set to `Git`.
The ktlint config is updated in the `app/build.gradle.kts` file.